Remove sbt-native-packager
From the history depending on this comes from the very first commit
(26ce7c5d0903eea70de34d5a775279bc4a9ef810) but AFAICT it didn't make use
of it in that commit.  And it doesn't make use of it now, either (again,
AFAICT).  So let's get rid of it.